SkinAtlas: cast LayerMask.value through uint before Mathf.Log
LayerMask default of (1 << 31) is int.MinValue (negative because bit 31 is the sign bit). Mathf.Log of a negative returns NaN, and Mono casts NaN to int as Int32.MinValue — far outside [0..31], triggering "layer needs to be in range" errors on gameObject.layer assignment.

Cast to uint first so bit 31 reads as 2147483648 and Log returns 31.

+++ Draw.ioTest/Draw.ioTest/Assets/Scripts/UI/SkinAtlas.cs
@@ -40,7 +40,11 @@ public class SkinAtlas : MonoBehaviour
         m_Camera.orthographicSize = Rows * m_CellWorldSize * 0.5f;
         m_Camera.aspect           = (float)m_Columns / Rows;
 
-        int layer = (int)Mathf.Log(m_Layer.value, 2);
+        // Cast through uint: m_Layer.value with bit 31 set is int.MinValue
+        // (negative), and Mathf.Log of a negative is NaN — which Mono casts
+        // to Int32.MinValue. As uint, bit 31 reads as 2147483648 and Log
+        // gives 31 correctly.
+        int layer = (int)Mathf.Log((uint)m_Layer.value, 2);
 
         for (int i = 0; i < count; i++)
         {
